udp 冗余包可配置

+++ contrib/tkcp/Packet.h
@@ -20,7 +20,7 @@ namespace packet {
 namespace tcp {
 
 
-const size_t kPacketHeadLength = 3;
+const int kPacketHeadLength = 3;
 
 struct TcpPacketHead {
     uint16_t len; // 包含包头长度
@@ -51,7 +51,7 @@ string  PacketIdToString(uint8_t packetId);
 
 namespace udp {
 
-const size_t kPacketHeadLength = 5;
+const int kPacketHeadLength = 5;
 struct UdpPacketHead {
     uint32_t conv;
     uint8_t  packetId;
